% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user_data.R
\name{create_user_data}
\alias{create_user_data}
\title{Create User Content Data}
\usage{
create_user_data(
  data = NULL,
  vars = NULL,
  save = T,
  path = NULL,
  file_type = "csv",
  fuzzy = T
)
}
\arguments{
\item{data}{Data from the BES Internet Panel, 2014-2023}

\item{vars}{A vector of character strings corresponding to the variable names for the given user content.}

\item{save}{Whether to save the data to the disk. Takes the values TRUE or FALSE.}

\item{path}{A file path as a character string specifying where to save the data if save == T.}

\item{file_type}{A character string specifying what file_type to save the data in. Takes the values "csv", "sav", or "dta". Defaults to "csv.}

\item{fuzzy}{If true, the function uses partial string matching. If false, it requires the complete string.}
}
\value{
Depending on the input file_type, either a data frame or tibble of data.
}
\description{
This function takes a copy of the full BES Internet Panel and returns a new data set indexed by respondent id that includes only a subset of variables.
}
\examples{
dta <-
  read_spss("BES2019_W19_Panel_v0.2.sav") \%>\%
  create_user_data(vars = c("user1", "user2"))
}
